PredrTest now handles arbitrarily long rules
We have a unit test in place that seems to confirm this is all working.
However, when we try to generate data we get a loop.  Could this be
a legitimate loop?  I suspect a bug somewhere.

Another potential medium-term next step is to put MaRz into Predr
when the rules have no proposed action sequence.