Added .laneidea and .nuxidea folders
The .idea folder contains configuration files that are are computer/OS
specific.  So, each time someone checks out the source they need to
be updated.  I've created a copy of what's there now (which works
on Braeden's computer) into a folder named .laneidea.  I've also created
a .nuxidea folder so that I can quickly replace the .idea folder
with one that works on my computer.