Unit Tests Working
I fixed the broken unit tests and revised the unit test code to remove a lot of the copy/paste that was going on.  I also observed unit tests need more work and more of them.  I'll keep gnawing on it.

To support the unit tests and debugging I improved the TreeNode.printTree() method family.

Some other comments and minor restructuring is in this revision as well.