Final(?) code overhaul
I've now reviewed and polished the entire agent as well as its
unit tests.

As I've worked, I've built up a TODO list in the comment header
for the PhuJusAgent class that we should use from now on (although
I imagine GitHub's issue tracking might be a more sophisticated option).