Installed Chaining Rewards Algorithm
The agent now gives a reward for a rule that correclty predicts a GOAL and then
chains backward giving discounted rewards to rules that led to that prediction.
It seems to be working and rule #1 is consistently being retained.  I've left
some extra printlns in the code for now for studying this.

I found that rules were getting ejected to soon after being added.  To combat
this I introduced a mechanism where a rule is removed every N rule firings where
N is a constant in the agent.

There is an occasional crash in addInternalCondition() I just found and don't
have time to diagnose.  To see it, remove the seed from rand and run a few times
until it manifests.

I moved intSensorInUse from Rule to PJA where it ought to have been anyway.

I fixed a bug in how paths are handled in getNextAction().

I fixed a bug in findBestPath().  It was finding the worst path...opps.