Minor tweaks to the FSM environment code
I was trying to get "new state:" messages not to spam at the end
of each run but it didn't work.  I'll try again later.

I also added some code from an old agent that calc's universal
sequence.  It needs to be ported but I have it there for now as
a template.