Universal Sequence Support
The FSM env now calculates a universal sequence as well as the shortest possible sequence to goal whenever the agent enters a new state.  This info is printed to output right now.

I've modified the data output to include (hard-coded) the average steps to g oal for a 6,2 FSM using the universal sequence to get there (blind path).  This will be our baseline for improving the agent for now.  Someday I'd like to have it calculate this average but the code is complicated and I ran out of steam trying to figure it out.