SoftMax Prediction, PathRules, Curiosity, Oh my!
I made a lot of changes here that should have been in separate
checkins:

Big Change:  Added PathRule class.  A PathRule describes a pattern at a higher scope than an EpRule.  Specifically it's LHS is (external sensors + path), (external sensors + path) and the RHS is a set of resulting external sensors.  Note that internal sensors are absent here.  I'm not sure that's the right move but I think it is.  The agent builds PathRules whenever it completes two paths in a row (whether goal is reached or not).  Then, when building a new path, a PathRule assessment is done a meta-score is factored in that is based on the PathRules' "opinion" of the path (see PJA.metaScorePath()).  In the end what this does is let the agent "notice" when it's in a loop and bail out right away.  However, the agent is still getting into loops sometimes.  We may need layers of rules to combat this.

Big Change: Added a voting RHS prediction algorithm wherein all matching rules vote on the outcome of a TreeNode + action.  I had high hopes for this (and still do!) but it's creating infinite loops right now even with the PathNodes.  I need to work on it.  It's currently turned off in favor of an alternative approach:  SoftMax RHS preduction.  This uses a SoftMax selection to select a single rule that will guess what happens next.  It's neat but I don't like it.

Medium Change: I added curiosity via "uncertain" paths.  When the agent can't find a path, it examines the tree and finds the shortest node it is least certain of: (1/pathlen)*inverse certainty and guesses it might be a goal.  This causes the agent to retry uncertain sequence and, in particular, try sequences it has never tried before.  This gets the agent a bigger breadth of experience sooner.

Small(er) Changes:
- added hashCode() overrids to the Conditions (bug fix for duplicate conditions)
- implemented overall rule accuracy again (improve rule set maintenance)
- EpRule.compareTo() requires actions to match (bug fix)
- expanded EpRule.updateConfidencesForPrediction
- moved reevaluateInternalSensors into EpRule to support the new child Path ctor
- removed the updateConfidencesForGood/BadPath and updateConfidencesForGoal() methods.  The PathRules make these obsolete.
- Both voting and softmax now also assign a confidence to a given path.  So the agent can weigh whether or not to try a path based upon (1/pathlen)*confidence.  I'm not sure this helps at all but I think it does.