lhsInternalMatch is a disjunction on a per-level basis
Consider this pair of rules:
      #15: |01b -> 00
      #25: (15)|00a -> 10

Now imagine the agent has a new experience that yields this candidate rule:
      #53: |01b -> 00

Since #15 and #25 have the same LHS they have to be resolved.  The first choice for doign this is to expand them yielding a result like this:

      #15: (8)|01b -> 00
      #53: (52)|01b -> 00

In this situation, consider rule #25.  The condition on its LHS was based upon #15 having a timedepth of 0.  Now its timedepth is 1 and therefore this revised #15 won't fire in all the same situations as could when #25 was created.

Proposed fix:  Rule Splitting

- positive conditions in the internal LHS of one level of a rule become a disjunction (aka, only one of them needs to match for that level to match).  Example:
     This rule:           (11)(4,5,23)(14,19)00a->00
     matches this state:  (11)(5)(19)

- when a rule A is expanded to resolve a conflict with rule B where the rules have matching LHS and matching RHS, any rule C that tests A on its LHS in any level now tests A or B (and vice versa).

This commit contains the first part of the above proposal.  It's important to note that the not-sensors are still conjunctive.

Other, smaller, changes in this revision:
- EpRule.removerIntSensor() has a return code again and PJA uses that to attempt to reintegrate rules that were truncated by the process.  It could try harder but it would be nasty code I don't want to write atm.  I'm not sure we need it anyway.
- Added EpRule.isEmptyLevel()  (and using it in places)
- refactored two helper methods out of updateRuleSet() so they can also be used to reintegreate truncated rules as mentioned above.  Specifically, added:
  . PJA.findBestMatchingRule()
  . PJA.resolveRuleConflict()
  . the RuleMatchProfile inner class which they share
- I've left in some commented out code to reward rules based on their internal sensors again.  I'm going to take it out but I don't want to lose it because I'm not 100% convinced it's wrong.