Rule resolving is working
Right now when the new rule is an exact match (LHS+RHS) to an  existing rule I have the code just ignoring the new rule.  There is a merge() method that needs to be implemented to merge them together.  I think that's the right thing to do.

I also need to go through the code and cull old content.