Got the Random numbers to seed consistently
Before now the behavior of the agent would change each run even if you used a random number seed.  I've removed the variable seed content of the framework to prevent this from happening.  It's kind of a nuclear option but I will go back and do it better if the need arises.