Removed the TF comparision in the Cond.equals() method.
I'm not 100% convinced but I don't think we should do this.
ALso made some minor edits.  In this session we discovered that TFRule.helperMatch should be comparing ext sensors values as well as names.  It was only doing the latter.  That's the next fix that should appear in this repository.