findMostUncertainPath looks 1 step ahead
In the longer term maybe needs to look N steps ahead again.  THe commented-out code is still there but I think it was causing loops.  I wonder if we should reconsider entirely what to do when no goal path is found and/or how we calculate path confidence.

Other changes
- changed NO_TFIDF to TFIDF (Nux hates double negative)
- made the RuleLoader parsing more robust to handle Nux's obsessive use of spaces