Fixed bug in genNextInternal
The method was using isExtMatch to verify which rules/internal sensors
should be on for the current time step.  This works fine for the agent
when setting the internal sensors for itself but it was the wrong
choice when the TreeNode ctor was trying to predict the internal sensors
of a child node of a given parent.

So we fixed this by having isExtMatch() require it be given the
sensors to match against and then creating two versions of genNextInternal:
one for the PJA and one for the TreeNode ctor.