TFRule's toString now shows internal sensor names
Instead of binaries (0s and 1s for internal sensors that are off/on respectively), the toString output now shows the names of the sensors that are on.

Also added an overloaded isMatch() method in TFRule which takes booleans for the different sensors instead of a TFRule object. Might be useful for debugging?