Initial Start on a new PathRule
Right now, PathRules are still not being used but I've revised the
class so it is a subclass or Rule and that it has this structure;
      rule + rule -> ext sensors.
Thus it can be used to create a hierarchy of rules.

I've added/edited/deleted instance variables and methods in PJA to
support the new PathRules.

I've also added a toStringShort() method to Rule because it serves the
toString method in PathRule.