Fixed Bug: Duplicate PathRules
The pathRule list in the PJA was getting duplicate rules.  I re-architected updatePathRules() to adjust this and also make sure a new PathRule obj isn't created just to throw it away.

To help with debugging the above, I adjusted the output of PathRule.toString to be brief and added (untested) support for the case where a LHS element is a PathRule.