Implemented Another New PathRule Idea
They are a lot more like TFRules.  In particular:

1.  We create them each time the agent completes a path and they are based on that path.

2.  We use the time depth paradigm as TF Rules.  TF Rules look like this:

        time depth 1:  (*)00a->00
        time depth 2+: (11)|00a->00

    Now PathRules should look like this:

        depth 1:  (null)|abaa->00
        depth 2+:   (11)|abaa->00

The "internal sensor" of the PathRule is the id of another PathRule.
The "external sensor" of a PathRule is the path
The RHS is the same.

For coding simplicity I actually implemented this as:
     public class PathRule extends Rule {

         private final PathRule lhs;  //can be 'null'
         private final Vector<TreeNode> rhs;

So, the "external sensors" are actually in the 'rhs' instance variable.