Fast and Loose PathRule Matching
Each PathRule has a "flat" instance variable that describes it
succintly as a series of actions (e.g., "abbab").  I use this to
(greatly!) speed up the matching at the cost of some loss of accuracy.

I also changed the code that sets prevPRMatch to only use the longest
matching PathRules instead of all matching PathRules.  This greatly
decreased the size of the PathRules.

I've added some additional debug statements to help with PathRule
performance monitoring.