Agent Rejects Paths that are Worse than Random
I added the final external sensors to the end of each PathRule's flat.  See the big comment on this as an explanation of why it's a temporary fix at best.

I had the agent reject paths that have a chance of success less than the success rate of random actions.