TF-IDF calculation now uses relevance
I've revised the calcualteTFIDF() method to include a relevance factor that is simply the abs(tf - df).  The closer tf and df are to each other, the less imporant the result (good or bad) should be.

Note that the match score is now in the range [-1.0..1.0] instead of [0.0..1.0].  I think we have to do this since there are effectively three poles in a tfidf value now: good, bad, irrelevant.

This change in match score range has these implications to the current code base:

1.  It's easy to set the internal match score for a wildcard rule: 0.0.  This makes a wildcard rule's overall score effectively just their ext LHS match which makes a lot of sense to me.

2.  The MIN_MATCH and MAX_MATCH hyper-params are no longer needed

3.  the PJA.calcNextInternal() method has to take into account that the bestScore maybe less than zero

4.  PJA.updateTFRuleConfidences() no longer relies on MATCH_NEAR hyper-param and is much simpler.

5.  We come head to head with an issue we've been dancing around: how much to weigh the internal match score vs internal.  Do they both get equal weight?  I've decided the overall score is the sum of the two but capped at 1.0.  This effectively gives them equal weight.  In the long term, weighting based on experience would eliminate this implicit hyper-parameter.

Other Changes:
 - I've been using metaScorePath() on the result of the call to fbgpHelper().  I've been studying some loops and I'm thinking it needs to be applied within fbgpHelper as soon as a path is scored.  It's more expensive that way but I think it's right.
 - I've fixed a bug in the code that tracks the random path success rate.