Caching DF Values
I ran some ad hoc profiing and it quickly became clear that the agent is
spending 84% of its time in TFRule.lhsIntMatchScore().  I found I could get a
big speedup in this method by caching the DF values (PJA.internalPercents).
So that's the major change in this revision.

Other changes:
- PathRule.addExample() now caps the length of positive and negative to 10.
  I did this just to delay the time when the agent runs out of heap.  I don't
  think it has any significant impact on the agent.
- Cond class now has a sensor id (int) in addition to a name.  This allows
  the cache (above) to run faster.
- revamped the code for lhsIntMatchScore()