Search Tree Favors Never-Tried Sequences
If the search tree encounters a path it's never tried before, it regards it as "high confidence" to encourage trying it out.  This seems to improve overall performance.

This revision also contains a change to PJA.updateTFRuleConfidences().  It is now more aggresssive about only adjusting rules that show strong evidence of matching.  This seems to improve performance on difficulty FSMs and makes the agent complete them faster.  I don't see any evidence of improvement in overall performance.