Profiling and Tuning
I moved MATCH_NEAR with the other hyper-parameters.  Some investigation shows that there is no good value for the variable so I left it as is.

I added code to lhsIntMatchScore() to penalize rules that don't have matches for the given sensors.  This will prevent rules like this:
  ()00a->10
from getting a higher match score than rules like this:
  (33)00a->10
when sensor 33 is on which seems important.

I did some more profiling and found that the biggest bottleneck remains in lhsIntMatchScore() which, of course, is now worse due to the change above.  I also found the recently-added hasOneLHSMatch() is a bottleneck as well.  I made some effort to minimize the runtime of these methods.