First, Buggy Impl of TimeDepth in TFRules
Okay, I've implemeted timeDepth in TFRules.  Now all TFRules have a primary internal sensor that must match for the TFRule to match.  The other internal sensors serve as "tie breakers."  This forces the search tree to follow legit sequences of episdoes when trying to find a goal.

This was a comprehensive change to the code and it's clear that bugs remain.  But it's stable enough I'm checking it in.  The TFRules are working and appear to be correct.