Multi-Depth Root TreeNode
An issue I've seen comes about when the agent has just reached a goal:
  Completed Episode: (109, 4, 215, 84)|00a -> 01
Episode in Progress: (115, 117, 116)  |01? ->

Because it's just reached a goal, the agent builds a new search tree.  The parent node of a tree currently has a depth of 0.  So, only the depth 0 rules are used.  This means that all those matching internal sensors are wasted.  In our current environment, this is okay since it's a random reset.  But you could imagine it not being so.  In this case, the internal sensors would be invaluable.

I modified TreeNode so that it's initial depth was "unknown" (-1) instead of 0.  This triggered the agent to search using rules at all levels on parent nodes.  In other words, this let the agent incorporate sequences that have GOAL=true in the middle.  I ran this and saw an improvement performance which is a bit surprising.