CondSet has base + confidences
Problem:  As a rule's Cond values change, it may end up in the wrong
bin (RuleIndex)!

Solution: I changed CondSet to be a set of fixed values (a base) and a
confidence in each value. Thus the conditions of a CondSet can't ever
flip BUT the confidence can drop to zero.  This makes more sense to me.