PathRules being used by NdxrAgent
Sort of.  There are bugs so it's stuck in a loop right now.

I've simplified how PathRules are used as follows (for now):
PathRules are only used to describe a path to the goal.  As a result, PathRules will be created when the agent creates a new path to the goal, a PathRule should be created based on that path (unless it already exists).  The LHS of the rule will be empty unless the agent has just reached the goal.  This is because the LHS is a PathRule and, per this design, a PathRule is only a path to the goal.

If the agent selects a random action that is treated as selecting new, one-step path and a PathRule is created for it.

The agent has an instance variable that stores the PathRule that the agent is currently following.  Thus, when a new PathRule is created the instance var can be used as its LHS.  Whenever the agent fails to reach the goal with a path, this variable is set to null so that new PathRules have an empty LHS.