created Conf class
I split out the confidence mechanism of CondSet into its own class, Conf, so that I could use it the track PathRule confidence instead of with a success rate.  This prevent infinite loops caused be the fact that the success rate would get in a race to the bottom with the agent's random success rate.  Now, it drops to complete zero when it gets low.

As a result of this change, the agent now completes a QUICK run without crashing. (yay!)  And it shows signs of learning! (YAY!)  More investigation needed.