New rules are only created if no perfect match for LHS or RHS
Until now only a perfect LHS match prevented a new rule from being created.  Now, if that fails. a perfect RHS is looked for.  If the LHS match was good and it predicted well then there is no need for a new rule.

Look particularly at NDXR.findBestRHSMatch() and how it is used for details.