Added more communication from Java to Python
The Java side sends the past 10 steps and current
sensor data to the Python side. There is an issue
of the alphabet being sent at the same time as the
history which causes them to be received in the same message.
The issue seems to only happen when a new FSM is created.