Created port_numbers log file
Note that the portNums log file needs to be deleted before running make runexperiments again because it is being run in the python side. TODO: write to portnums file from the java side of the socket and empty the file before running.

This was to catch a race condition error where python agent port numbers were being reused causing issues the program to freeze.