Fixing a bug where the agent gets stuck in a loop
The beginning of code that is trying to get the
agent out of an infinite loop. The plan is to
check if the number of steps since the last goal
exceeds a certain amonut, if so, revert back to
random choices. When new goal is reached, create
and train a new model.