Completed model simulations
After a model is trained it is then used to generate a predicted path (max length=50) to the next goal.

calc_desired_actions() in TFSocketModel.py was modified to manage not only 'a' and 'b' but 'A' and 'B' (or the 'reached goal state' equivalents) as well. This solution has cause inaccuracies in the predictions and possibly other issues with the model's accuracy.