Adjusted code for general alphabet and made model changes
Made changes so that all the places that were hardcoded for a two-letter
alphabet to are now able to handle an alphabet of any size. This current
version doesn't seem to be functioning, the sim_path is always 'A'. Also
changed the __init__ call for tfmodel() by taking out the model
initialization there and putting it in train_model instead. Still need to test how this affects
performance.