More progress, ran into a big issue
Added more support for sensors. Currently,
flatten() fully works when used in calc_input_tensors()
and calc_desired_actions() works (assuming it doesn't need
sensor data). The big issue is that in simulate_model(), we
use the entire_history, but it now contains sensor data, which
the model isn't designed to predict. Not sure how to proceed from this point
but an idea is to have two separate flattens, one used for training data
and the other used for simulating.