Beginning PyTorch DQN Model
socket communication with the java side
and the pytorchDQN model is working. Agent
is at the state of only sending a random letter
- we should consider adding print statements for
the agent's actions back in.

NOTE: Can ignore the tfsocket data changes, think I was
halfway through updating the repo data but never got around
to finishing that task.