learning/unlearning sigmoid mapping
The program agent is taking the avg_steps_to_goal of the last 5 goals and check if it is greater than the avg_steps_to_goal of the total history. Then define an increasing sigmoid function with the lower bound = the current epsilon at which the agent begun "unlearning" with the upper bound =1. Then when the agent starts learning again it redefines the sigmoid function with the upper bound = the current epsilon, and the lower bound = 0. This method is what seemed intuitive to me after doing some investigating and brainstorming.