bug fixing
Note: the 1 model only trains once at the model creation threshold because the looping detector is not working in main, see branch "revised-looping-threshold"

bug fix 1.
sim_paths were not being updated when the agent took a random step for getting out of a loop (a similar fix will be necessary for the egreedy branch as well)

bug fix 2.
in process_history_sentinel() when it was processing that the last step the agent made found a goal, the agent was returning before moving on in the function to the model portion of the function. This effectively cause the agent to repeat its last step taken after finding a goal, as opposed to the AI agent informed predicted best step.