Nuxoll makes more re-arch moves
We are in the process or getting Yuji's code to talk to our framework.
We ran into trouble with self.env which was a DFAEnv object but we
don't want to use that class anymore because the framework is our
env now.  So, some pretty deep changes were made to accommodate this
including a new logReward() method on the .py side.

The work isn't complete.  I've added TODO items that hopefully
will get you most of the way there.